fix sneaky bug after rewriting recalcEnergy to use copyWithin
This bug happens because of a difference between the old and new
copyWithin functions, the new one copies the entire line *before*
recalculating the energy of pixels around the seam.

Adding the offset gets you to the correct pixel in the *old* row. Now
the row has already had the seam removed, the offset doesn't need to be
added.